pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "galactic-neighbors-finder"
version = "1.0.0"
description = "Efficient neighbor identification in galaxy catalogs using k-d trees"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Deepak Deo", email = "deepak.deo@example.com"}
]
keywords = ["astronomy", "galaxies", "neighbors", "kdtree", "cosmology"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Astronomy",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "numpy>=1.19.0",
    "pandas>=1.0.0",
    "astropy>=4.0",
    "scipy>=1.5.0",
    "pyyaml>=5.3",
    "tqdm>=4.50.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.10",
    "black>=20.8b1",
    "flake8>=3.8",
    "isort>=5.0",
    "mypy>=0.900",
]

[project.urls]
Homepage = "https://github.com/deepakdeo/GalacticNeighborsFinder"
Documentation = "https://github.com/deepakdeo/GalacticNeighborsFinder"
Repository = "https://github.com/deepakdeo/GalacticNeighborsFinder.git"

[project.scripts]
gnf-finder = "gnf.cli:main"

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--cov=gnf --cov-report=term-missing"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
